requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
email-validator==1.3.0
fhir.resources==6.4.0
idna==3.4
pydantic==1.10.2
requests==2.28.1
typing_extensions==4.3.0
urllib3==1.26.12
